feat: implement automark pipeline
